chassis: add 'creator' flag for clearing env and fix wrong usage of member variable

+++ infrasim/chassis/share_memory.py
@@ -11,8 +11,11 @@ import posix_ipc
 
 
 class CShareMemory:
-    handle_memory = None
-    handle_file = None
+
+    def __init__(self):
+        self.handle_memory = None
+        self.handle_file = None
+        self.is_creator = False
 
     def create(self, key_name, size):
         if self.handle_memory is not None:
@@ -22,6 +25,7 @@ class CShareMemory:
         flags = posix_ipc.O_CREAT  # posix_ipc.O_CREX
         self.handle_memory = posix_ipc.SharedMemory(key_name, flags, mode=0o644, size=size)
         self.handle_file = mmap.mmap(self.handle_memory.fd, self.handle_memory.size)
+        self.is_creator = True
         return self.handle_file
 
     def open(self, key_name):
@@ -40,7 +44,8 @@ class CShareMemory:
         if self.handle_file is None:
             raise Exception()
         self.handle_file.close()
-        self.handle_memory.unlink()
+        if self.is_creator:
+            self.handle_memory.unlink()
         self.handle_memory.close_fd()
 
     def write(self, position, src):
